bandwidth: wait on map event
Some implementations (e.g. Clover) return from the clEnqueueMap* API
calls before the event is completed. This makes the profiling info not
available to the subsequent peformance queries. Fix by waiting on the
map event before getting the profiling info (due to the waiting lists,
this should be enough to ensure that all events have completed).